Update version and refactor upload method
The version number in the pom.xml and build.sh files has been updated to 0.0.5. Also, the upload and download methods in the PerfanaConfigGeneratorController.kt have been refactored, merging them into a single method named generate. The function of fetching, generating, and returning the zip resources are now handled by the "generate" method.